+++ src/core/index.ts
@@ -1,2 +1,3 @@
 export * from "./core.js";
+export { applyElement } from "./jsx/common.js";
 export * from "./view-jsx.js";